VARS-749: RS422 always uses VCR.requestTimecode. Using VCR4j > 2.0.1 with fixed UB commands

+++ vars-annotation/src/main/java/vars/annotation/ui/video/UDPVideoControlService.java
@@ -53,7 +53,7 @@ public class UDPVideoControlService extends AbstractVideoControlService {
 
         String host = (String) args[0];
         Integer port = (Integer) args[1];
-        frameRate = ((Double) args[2]).doubleValue();
+        frameRate = (Double) args[2];
 
         try {
             final IVCR vcrUdp = new VCR(host, port);
@@ -104,7 +104,6 @@ public class UDPVideoControlService extends AbstractVideoControlService {
         final Timecode timecode = vcrTimecode.getTimecode();
         final Timecode updatedTimecode = new Timecode(timecode.getFrames(), frameRate);
         vcrTimecode.timecodeProperty().set(updatedTimecode);
-        //timecode.setFrameRate(frameRate);
 
         // Convert userbits from byte[]->long->Date
         final int epicSeconds = NumberUtilities.toInt(vcrUserbits.getUserbits(), true);
